+++ api/responses/response_wrapper.py
@@ -1,0 +1,5 @@
+def paginated_response(data, total_count: int):
+    return {
+        "data": data,
+        "total_count": total_count,
+    }