+++ tests/test_base_repository.py
@@ -1,12 +1,11 @@
 import pytest
-from sqlalchemy.exc import IntegrityError
-
 from database.models import TyreModel
+from sqlalchemy.exc import IntegrityError
 from database.repositories.base_repository import BaseRepository
 
 
-def test_create(database_session):
-    repo = BaseRepository(database_session, TyreModel)
+def test_create(app_ctx, database_session):
+    repo = BaseRepository(TyreModel)
 
     # Can not create entity with required fields missing
     with pytest.raises(IntegrityError):
@@ -20,8 +19,46 @@ def test_create(database_session):
     assert tyre_model.model_name is not None
     assert tyre_model.model_name == 'Test Tyre Model'
 
-def test_get_by_id(database_session):
-    repo = BaseRepository(database_session, TyreModel)
+def test_get_all(app_ctx, database_session):
+    repo = BaseRepository(TyreModel)
+
+    tyre_model_1 = repo.create(manufacturer='Michelin', model_name='Test Tyre Model')
+    tyre_model_2 = repo.create(manufacturer='Not Michelin', model_name='Different Model Name')
+    tyre_model_3 = repo.create(manufacturer='Pirelli', model_name='Third Model')
+
+    # Can get all results with no pagination
+    results, total_count = repo.get_all()
+
+    assert len(results) == 3
+    assert total_count == 3
+    assert results[0].id == tyre_model_1.id
+    assert results[1].id == tyre_model_2.id
+    assert results[2].id == tyre_model_3.id
+
+    # Can limit number of results returned with pagination
+    limit_results, total_count = repo.get_all(limit=2)
+
+    assert len(limit_results) == 2
+    assert total_count == 3
+    assert limit_results[0].id == tyre_model_1.id
+    assert limit_results[1].id == tyre_model_2.id
+
+    # Can offset start point of pagination
+    offset_results, total_count = repo.get_all(limit=2, offset=1)
+
+    assert len(offset_results) == 2
+    assert total_count == 3
+    assert offset_results[0].id == tyre_model_2.id
+    assert offset_results[1].id == tyre_model_3.id
+
+    # Can not return results with invalid offset
+    empty_results, total_count = repo.get_all(offset=10)
+
+    assert len(empty_results) == 0
+    assert total_count == 3
+
+def test_get_by_id(app_ctx, database_session):
+    repo = BaseRepository(TyreModel)
 
     # Can get by ID
     created = repo.create(manufacturer='Michelin', model_name='Test Tyre Model')
@@ -40,47 +77,9 @@ def test_get_by_id(database_session):
 
     assert found is None
 
-def test_get_all(database_session):
-    repo = BaseRepository(database_session, TyreModel)
 
-    tyre_model_1 = repo.create(manufacturer='Michelin', model_name='Test Tyre Model')
-    tyre_model_2 = repo.create(manufacturer='Not Michelin', model_name='Different Model Name')
-    tyre_model_3 = repo.create(manufacturer='Pirelli', model_name='Third Model')
-
-    # Can get all results with no pagination
-    results, count = repo.get_all()
-
-    assert len(results) == 3
-    assert count == 3
-    assert results[0].id == tyre_model_1.id
-    assert results[1].id == tyre_model_2.id
-    assert results[2].id == tyre_model_3.id
-
-    # Can limit number of results returned with pagination
-    limit_results, count = repo.get_all(limit=2)
-
-    assert len(limit_results) == 2
-    assert count == 2
-    assert limit_results[0].id == tyre_model_1.id
-    assert limit_results[1].id == tyre_model_2.id
-
-    # Can offset start point of pagination
-    offset_results, count = repo.get_all(limit=2, offset=1)
-
-    assert len(offset_results) == 2
-    assert count == 2
-    assert offset_results[0].id == tyre_model_2.id
-    assert offset_results[1].id == tyre_model_3.id
-
-    # Can not return results with invalid offset
-    empty_results, count = repo.get_all(offset=10)
-
-    assert len(empty_results) == 0
-    assert count == 0
-
-
-def test_delete(database_session):
-    repo = BaseRepository(database_session, TyreModel)
+def test_delete(app_ctx, database_session):
+    repo = BaseRepository(TyreModel)
 
     # Can delete model that exists
     tyre_model = repo.create(manufacturer='Delete Me', model_name='Delete Me')