+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.0.3 `DATE` - Test Coverage: 
 ### Added
 - Added pagination limit and offset to base repo get_all method
+- Added `get_all` route for `tyre_model` and response objects
 ### Changed
 - Fixed `.env.dist` to match required environment variables
 - Omitted `database/session.py` and `database/extensions.py` from test coverage report